eepbot_py/pyproject.toml
[project]
name = "eepbot-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.4.0",
    "mistralai>=1.5.0",
    "openai>=1.61.0",
    "seqlog>=0.4.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.4",
]
